refactor(ui): promote shared field components to organisms
