+++ src/app/(admin)/admin/layout.test.tsx
@@ -19,10 +19,8 @@ vi.mock("@/ui/components/atoms", async (importOriginal) => {
   };
 });
 vi.mock("@/ui/components/organisms", () => ({
-  SidebarToggle: (): null => null,
-}));
-vi.mock("@/ui/components/molecules", () => ({
   SidebarNavItem: (): null => null,
+  SidebarToggle: (): null => null,
 }));
 vi.mock("./_components", () => ({ AdminModal: (): null => null }));
