+++ src/app/(admin)/admin/reviews/_components/AdminReviewsTemplate.tsx
@@ -4,7 +4,8 @@ import { useTransition } from "react";
 import { useRouter, usePathname } from "next/navigation";
 import { toast } from "sonner";
 import { Button, TypographyH1, TypographyMuted } from "@/ui/components/atoms";
-import { CursorPagination, RatingStars } from "@/ui/components/molecules";
+import { CursorPagination } from "@/ui/components/molecules";
+import { RatingStars } from "@/ui/components/organisms";
 import type { AdminReviewListPage } from "@/core/domain";
 import { formatKstDate } from "@/core/utils";
 import { deleteReviewByAdmin } from "@/actions";
