+++ src/app/(main)/(checkout)/payment/_components/BuyerInfoCard.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Card, CardContent, CardHeader, CardTitle } from "@/ui/components/atoms";
-import { TextField } from "@/ui/components/molecules";
+import { TextField } from "@/ui/components/organisms";
 import type { BuyerInfo } from "@/core/schemas";
 
 interface BuyerInfoCardProps {
