+++ src/ui/components/molecules/BankField.tsx
@@ -1,7 +1,8 @@
 "use client";
 import { Input } from "@/ui/components/atoms";
 import type { BanksResponse } from "@/core/schemas";
-import { FormField, BaseSelect } from "@/ui/components/molecules";
+import { BaseSelect } from "@/ui/components/molecules";
+import { FormField } from "@/ui/components/organisms";
 import { useState } from "react";
 
 interface BankAccountInfo {
