+++ src/ui/components/molecules/DateField.tsx
@@ -8,7 +8,7 @@ import { Button, Calendar, Popover, PopoverContent, PopoverTrigger, Input } from
 
 
 
-import { FormField } from "@/ui/components/molecules";
+import { FormField } from "@/ui/components/organisms";
 import type { FieldBase } from "@/core/domain";
 import { cn } from "@/core/utils";
