+++ src/ui/components/organisms/FindIdForm.tsx
@@ -4,7 +4,7 @@ import { Mail } from "lucide-react";
 import { Card, Button, TypographyH1, TypographyLarge, TypographyMuted } from "@/ui/components/atoms";
 
 
-import { TextField } from "@/ui/components/molecules";
+import { TextField } from "@/ui/components/organisms";
 import { getFieldError } from "@/core/utils";
 import type { APIResponse } from "@/core/domain";
 import { routes } from "@/core/domain";
