+++ src/ui/components/organisms/LoginForm.tsx
@@ -5,7 +5,7 @@ import { Button, TypographyH1, TypographyMuted, Checkbox, Label } from "@/ui/com
 
 
 
-import { TextField } from "@/ui/components/molecules";
+import { TextField } from "@/ui/components/organisms";
 import { routes } from "@/core/domain";
 interface LoginFormProps {
   action: (formData: FormData) => void;
