+++ src/ui/components/organisms/SelectField.tsx
@@ -1,7 +1,8 @@
 "use client";
 
 import type { SelectOption } from "@/ui/components/molecules";
-import { FormField, BaseSelect } from "@/ui/components/molecules";
+import { BaseSelect } from "@/ui/components/molecules";
+import { FormField } from "@/ui/components/organisms";
 import { useState } from "react";
 import type { FieldBase } from "@/core/domain";
